# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Home Fileshare
  description: LAN photo/video file server with per-user folder ACLs
  version: 1.0.0
paths:
...
  /api/v1/admin/users:
    get:
      tags:
      - admin
      summary: List Users
      operationId: list_users_api_v1_admin_users_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserInfo'
                type: array
                title: Response List Users Api V1 Admin Users Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - admin
      summary: Create User
      operationId: create_user_api_v1_admin_users_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/admin/grants:
    get:
      tags:
      - admin
      summary: List Grants
      operationId: list_grants_api_v1_admin_grants_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/GrantInfo'
                type: array
                title: Response List Grants Api V1 Admin Grants Get
      security:
      - HTTPBearer: []
    put:
      tags:
      - admin
      summary: Set Grant
      operationId: set_grant_api_v1_admin_grants_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrantRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/admin/grants/{grant_id}:
    delete:
      tags:
      - admin
      summary: Delete Grant
      operationId: delete_grant_api_v1_admin_grants__grant_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: grant_id
        in: path
        required: true
        schema:
          type: integer
          title: Grant Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    CreateUserRequest:
      properties:
        username:
          type: string
          maxLength: 64
          minLength: 2
          title: Username
        password:
          type: string
          maxLength: 128
          minLength: 6
          title: Password
        is_admin:
          type: boolean
          title: Is Admin
          default: false
      type: object
      required:
      - username
      - password
      title: CreateUserRequest
    GrantInfo:
      properties:
        id:
          type: integer
          title: Id
        username:
          type: string
          title: Username
        logical_path:
          type: string
          title: Logical Path
        physical_path:
          type: string
          title: Physical Path
        can_read:
          type: boolean
          title: Can Read
        can_write:
          type: boolean
          title: Can Write
      type: object
      required:
      - id
      - username
      - logical_path
      - physical_path
      - can_read
      - can_write
      title: GrantInfo
    GrantRequest:
      properties:
        username:
          type: string
          title: Username
        logical_path:
          type: string
          title: Logical Path
        physical_path:
          type: string
          title: Physical Path
        can_read:
          type: boolean
          title: Can Read
          default: true
        can_write:
          type: boolean
          title: Can Write
          default: false
      type: object
      required:
      - username
      - logical_path
      - physical_path
      title: GrantRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UserInfo:
      properties:
        id:
          type: integer
          title: Id
        username:
          type: string
          title: Username
        is_admin:
          type: boolean
          title: Is Admin
      type: object
      required:
      - id
      - username
      - is_admin
      title: UserInfo
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer